lab2_sessions_memory/pyproject.toml
[project]
name = "lab2-sessions-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
]
